Accept votes that are equal to current proposal.
Otherwise if all nodes agree on a leader, each one of them will have
only one or two accepted votes. If a node is the agreed (would be)
leader, it will not accept any votes from other nodes. Rest of the nodes
will accept leader's vote and its own. Hence, if number of nodes is
greater than two, it's impossible to achieve the quorum.

Change-Id: I2d1c98dfc21257ac15a2305920a5991f5006b1d6
Reviewed-on: http://review.couchbase.org/42500
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>