Wait grace period after reaching quorum of votes for a leader.
During the grace period new votes are still accepted. If during this
time our idea of leader doesn't change, only then we declare it a
leader. This is what zookeeper also does to reduce the probability for
certain node to elect an incorrect leader which then causes
follower/leader synchronization to fail.

Change-Id: Ifa8715c5468972711d7b4b9c61f02586e17263eb
Reviewed-on: http://review.couchbase.org/42947
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>