MB-13774: index list not populated for client after indexer restart
Each update to the dictionay is assigned with a txid.  When a client
connects to an indexer, it reads off a specific txid.  Upon crash,
the dictionary needs to create a snapshot off the last txid being
committed, such that the client can read from this txid.

the client can read off

Change-Id: Id3da40270b9b76339bd35b4d0fc021aa932d198a