MB-39698 Change forestDB commit flag to COMMIT_MANUAL_WAL_FLUSH
With COMMIT_NORMAL flag, during a commit to forestDB, the
write-ahead-log (WAL) may not get flushed. When creating a snaphot,
if the WAL is not flushed, then there is a special logic to make sure
that the WAL items are a part of the snapshot. This logic is CPU
intensive. With increase in number of indexes, the time spent in
this logic increases and therefore gometa updates take more time.

With COMMIT_MANUAL_WAL_FLUSH, the WAL items gets flushed and therefore
gometa updates scale well.

Change-Id: I895481170a82a5b0d526d3d23e95c04a7c2d191d