MB-55246 - adjust log level for watcher on valid closes
BP to 7.2.0

bg: we create metadata provider object on all indexer nodes periodically
this starts watcher for each node. upon cleanup, peer pipe logs an error
that connection is terminated but this is expected behaviour. this error
log is creating alerts in customer log monitors

exp: if we are killing the peer pipe connection on a valid condition
don't log it as error

stubs: in my experiments, when we kill a server (with -9/-15 flag)
the connection is still gracefully terminated and go returns io.EOF
because of this we cannot diff between indexer crash and conn close
at gometa level. metadata provider will still be aware if conn was
expected to die or closed abruptly

assumptions: -
Change-Id: I015ac0e036a1ad07ffacb20d3ef760149047d45d