MB-69723 Close peer connection if message is received on proxy killch and donech simultaneously
background:

If messages are received on both proxy.donech and  proxy.killch together and golang somehow reads message from killch first, ignoring the message in donch , killch returns without killing the peer.conn (tcp connection), which results in connection leak.

Follower will not form a new connection and carryon being connected with leaked connection which is removed for leader's bookkeeping causing communication breakdown between follower and leader.

Fix:

This fix closes connection even if message is received on both channels and killch gets prioritised.

Also if LeaderSyncProxy fails to synchronize with the follower, we need to close the connection, which was not happening before.

Change-Id: I6d9667f1480a15052c4cefc05255441cf0b87f5a